Split CUDADeviceManager.h and OMPDeviceManager.h into cpp files.
