+++ src/Platforms/OMPTarget/CMakeLists.txt
@@ -2,14 +2,14 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2021 QMCPACK developers.
+#// Copyright (c) 2023 QMCPACK developers.
 #//
 #// File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//
 #// File created by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//////////////////////////////////////////////////////////////////////////////////////
 
-set(OMP_RT_SRCS OMPallocator.cpp)
+set(OMP_RT_SRCS OMPallocator.cpp OMPDeviceManager.cpp)
 set(OMP_LA_SRCS ompBLAS.cpp)
 
 add_library(platform_omptarget_runtime ${OMP_RT_SRCS})
@@ -18,6 +18,7 @@ if(USE_OBJECT_TARGET)
 else()
   add_library(platform_omptarget_LA ${OMP_LA_SRCS})
 endif()
+target_link_libraries(platform_omptarget_runtime PRIVATE platform_host_runtime)
 target_link_libraries(platform_omptarget_LA PUBLIC platform_omptarget_runtime)
 
 target_compile_options(platform_omptarget_LA PRIVATE "$<$<BOOL:${ENABLE_OFFLOAD_CLANG_DEBUG_O3}>:$<$<CONFIG:DEBUG>:-O3>>")
